Create next.config.js if it is missing. (#701)
* Create next.config.js if it is missing.

* Add comment.

+++ stack_orchestrator/data/container-build/cerc-nextjs-base/scripts/build-app.sh
@@ -21,6 +21,11 @@ WORK_DIR="${1:-/app}"
 
 cd "${WORK_DIR}" || exit 1
 
+# If this file doesn't exist at all, we'll get errors below.
+if [ ! -f "next.config.js" ]; then
+  touch next.config.js
+fi
+
 if [ ! -f "next.config.dist" ]; then
   cp next.config.js next.config.dist
 fi
